+++ IoTDevices/custom_components/fusb302/UsbPdStateMachine.h
@@ -32,8 +32,8 @@ public:
   // Gets the capabilities of the source. Returns the total count, and the unpacked capabilities are
   // stored in the input array.
   // Zero means no source is connected, or capabilities are not yet available.
-  // Can return up to 8 objects.
-  int getCapabilities(UsbPd::Capability::Unpacked capabilities[]);
+  // Can return up to kMaxCapabilities=8 objects.
+  uint8_t getCapabilities(UsbPd::Capability::Unpacked capabilities[]);
 
   // The currently active capability requested to and confirmed by the source.
   // Zero means the default (none was requested).
